feat(admin): create new component InformationEdit
move the edit of the information into the new component InformationEdit
